Use size_t and cast appropriately

+++ src/transformer/manager.cpp
@@ -56,7 +56,7 @@ bool manager::transform(const ddwaf_object &source, ddwaf_object &destination,
     }
 
     bool transformed = false;
-    lazy_string str({source.stringValue, source.nbEntries});
+    lazy_string str({source.stringValue, static_cast<std::size_t>(source.nbEntries)});
     for (auto transformer : transformers) { transformed |= call_transformer(transformer, str); }
 
     if (!transformed || !str.modified()) {
